BVTCK-89 Split BuiltinConstraintsTest to improve readability
